finish v0.1

+++ ords/setup_ords.sh
@@ -1,11 +1,21 @@
 #!/bin/bash
 
-# catalina.sh stop
-
-cd /opt
-ords install --db-hostname ${ORACLE_HOST} --db-port ${ORACLE_PORT} --db-servicename XEPDB1 --feature-sdw true --admin-user sys --password-stdin << EOF
+if [ -f ${ORDS_CONFIG}/.ords.created ]; then
+    sts=10
+    echo File .ords.created sudah ada !
+else
+    cd /opt
+    ords install --db-hostname ${ORACLE_HOST} --db-port ${ORACLE_PORT} --db-servicename XEPDB1 --feature-sdw true --admin-user sys --password-stdin << EOF
 ${ORACLE_PWD}
 EOF
+
+    sts=$?
+fi
+
+echo Status $sts akan setup flag file .ords.created
+if [ $sts -eq 0 ]; then
+    echo "$(date -Iseconds)" > ${ORDS_CONFIG}/.ords.created
+fi
 
 echo Akan restart tomcat
 catalina.sh run
